[tidying up] splitted basic draw function into their own files separated from advanced display code [draw] added draw_horizontal/vertical_line functions
